+++ src/gui/controls/ClientListModel.cpp
@@ -21,10 +21,10 @@ QString formatSize(int64_t bytes)
     if (bytes < 1024)
         return QStringLiteral("%1 B").arg(bytes);
     if (bytes < 1024 * 1024)
-        return QStringLiteral("%1 KiB").arg(bytes / 1024.0, 0, 'f', 1);
+        return QStringLiteral("%1 KiB").arg(static_cast<double>(bytes) / 1024.0, 0, 'f', 1);
     if (bytes < 1024LL * 1024 * 1024)
-        return QStringLiteral("%1 MiB").arg(bytes / (1024.0 * 1024.0), 0, 'f', 1);
-    return QStringLiteral("%1 GiB").arg(bytes / (1024.0 * 1024.0 * 1024.0), 0, 'f', 2);
+        return QStringLiteral("%1 MiB").arg(static_cast<double>(bytes) / (1024.0 * 1024.0), 0, 'f', 1);
+    return QStringLiteral("%1 GiB").arg(static_cast<double>(bytes) / (1024.0 * 1024.0 * 1024.0), 0, 'f', 2);
 }
 
 /// Format a speed value.
@@ -34,7 +34,7 @@ QString formatSpeed(int64_t bytesPerSec)
         return {};
     if (bytesPerSec < 1024)
         return QStringLiteral("%1 B/s").arg(bytesPerSec);
-    return QStringLiteral("%1 KiB/s").arg(bytesPerSec / 1024.0, 0, 'f', 1);
+    return QStringLiteral("%1 KiB/s").arg(static_cast<double>(bytesPerSec) / 1024.0, 0, 'f', 1);
 }
 
 /// Format wait time as duration.
